+++ src/heapq/newHeapQ.c
@@ -1,6 +1,5 @@
 #include <errno.h>
 #include <manos.h>
-#include <stddef.h>
 
 /**
  * newHeapQ() - allocate memory for a priority heap
@@ -9,7 +8,7 @@
  *   pointer suitable for freeing with kfree
  */
 HeapQ* newHeapQ(size_t size) {
-    HeapQ* q = syskmalloc(sizeof(*q) + (sizeof(intptr_t) * size));
+    HeapQ* q = syskmalloc(sizeof(*q) + (sizeof(uint32_t) * size));
     if (!q) {
         errno = ENOMEM;
         return NULL;
